packetdrill: replace ADD_ADDR v0 with v1
This change replaces the v0 implementation of the ADD_ADDR option with
v1.  The syntax of the option adds an optional 'hmac=auto' or
'hmac=INTEGER' to the end of the existing syntax.  Use the 'auto'
value to have packetdrill compute the expected hmac or use the INTEGER
value to explicitly set it.

Internally, the ipver field has been replaced by flag_E and the new
permutations of the option's values have been added.

Signed-off-by: Todd Malsbary <todd.malsbary@intel.com>