packetdrill: enable ADD_ADDR IP values to be variables
This change replaces 'ipv4=inet_addr("...")' in the ADD_ADDR option
with 'addr[NAME=inet_addr("...")]' and 'addr[NAME]'.  The behavior is
the same as 'key[NAME=...]' and 'key[NAME]' in the MP_CAPABLE option:
the value provided or generated is remembered in mp_state.vars.

This lays the foundation for using these values in subsequent subflow
creation.

Signed-off-by: Todd Malsbary <todd.malsbary@intel.com>